refactor(conformance): share directive tsconfig conversion
